% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inhouse_calc.R
\name{inhouse_calc}
\alias{inhouse_calc}
\title{Inhouse Customer Score Calculation}
\usage{
inhouse_calc(data = inhouse_data)
}
\arguments{
\item{data}{Dataset; must have our predefined column names: customer_qual1 to customer_qual5.}
}
\value{
Vector of customer_score values, in a range between 0 and 100.
}
\description{
Calculating customer score from our 5 key customer qualities, weighted by importance.
}
\examples{
inhouse_data$customer_score <- inhouse_calc()
inhouse_data$customer_score <- inhouse_calc(data = inhouse_data)

}
