% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inhouse_data}
\alias{inhouse_data}
\title{Data to showcase our Customer Score Calculation}
\format{
A data frame with 100 rows and 5 variables:
\describe{
\item{customer_qual1}{1st secret customer quality}
\item{customer_qual2}{2nd secret customer quality}
\item{customer_qual3}{3rd secret customer quality}
\item{customer_qual4}{4th secret customer quality}
\item{customer_qual5}{5th secret customer quality}
}
}
\source{
{Created in-house to serve as an example.}
}
\usage{
inhouse_data
}
\description{
Contains the five secret customer qualities.
}
\examples{
data(inhouse_data)
}
\keyword{datasets}
